Render links in base template properly
Depending on whether the website is running on the top level or at
a nested level, some relative links may stop working. This only
affects base templates because they are used by pages on nested level.
Hence, create a relativize_link macro that checks the TEMPLATE_LEVEL
of the page and renders the appropriate relative link.